requirements.txt
langchain_community
google-cloud-aiplatform
langchain
pandas